fix(ui): Fixed items disappearing bug
The bug was caused by the missing `this@replace` when refering to `size` (which would therefore return a sublist of the size of the newly-created list, which is always empty).

A solution could be to simply add the missing qualifier, but instead I opted for a rewrite to make these errors impossible in the future (now only the old list is a receiver).