refactor(ui): Rewrote the field selection in the search bar
The previous implementation was hard to maintain and relied on multiple React renders to generate the full list of options (averaging 3–4 renders to correctly display the fields).

The new implementation generates a non-recursive list of options, and displays it in a single render.
The user can now see all available options in a single drop-down, which will hopefully make the search easier to use.