fix(ui): Empty FormData during Review
During review, action fields were missing a 'name' attribute. They therefore did not appear in FormData, and the system could not review any record with such fields.

This problem might have been signaled in https://gitlab.com/arcachon-ville/formulaide/-/issues/241 already, but the description is unclear. At this point, it is unknown whether this issue is solved by this commit or not.