perf(server): Block DoS attacks with too large file uploads
Previously, it was possible to upload a file that was an order of magnitude larger than the allowed maximum. The server would load the entire file in memory, and then would check that the size was not too big.

It was easy to upload a file much too big to saturate the server's RAM, and possibly crash it.

Closes https://gitlab.com/arcachon-ville/formulaide/-/issues/195