feat(server): The server attempts to create the default user on startup
In previous versions, the server had a particular system for initialization and creation of the first user.

This system is removed. Instead, the server always attempts to create the system administrator.

In the future, this should display an alert in the administrator interface.

See https://gitlab.com/opensavvy/formulaide/-/issues/47